+++ src/components/metadataFormControls/OptionSetSelect/index.ts
@@ -1,1 +1,0 @@
-export { OptionSetSelect } from './OptionSetSelect'