+++ src/lib/category/index.ts
@@ -1,1 +1,0 @@
-export { useDefaultCategoryComboQuery } from './useDefaultCategoryComboQuery'