+++ src/pages/dataElementGroups/form/index.ts
@@ -1,3 +1,6 @@
 export { DataElementGroupFormFields } from './DataElementGroupFormFields'
-export { dataElementGroupSchema } from './dataElementGroupSchema'
-export type { FormValues } from './types'
+export {
+    dataElementGroupFormSchema,
+    initialValues,
+    validate,
+} from './dataElementGroupSchema'