+++ src/pages/dataElements/form/dataElementSchema.ts
@@ -1,36 +1,75 @@
 import { z } from 'zod'
-import { modelFormSchemas } from '../../../lib'
+import {
+    createFormValidate,
+    DEFAULT_CATEGORY_COMBO,
+    getDefaults,
+    modelFormSchemas,
+} from '../../../lib'
+import { DataElement } from '../../../types/generated'
 
-const { identifiable } = modelFormSchemas
-export const dataElementSchema = identifiable.merge(
-    z
+const { identifiable, withAttributeValues, withDefaultListColumns } =
+    modelFormSchemas
+
+const dataElementBaseSchema = z.object({
+    shortName: z.string().trim(),
+    code: z.string().trim().optional(),
+    description: z.string().trim().optional(),
+    formName: z.string().trim().optional(),
+    url: z.string().trim().optional(),
+    fieldMask: z.string().trim().optional(),
+    style: z
         .object({
-            name: z.string().trim(),
-            shortName: z.string().trim(),
-            code: z.string().trim(),
-            description: z.string().trim(),
-            formName: z.string().trim(),
-            url: z.string().trim(),
-            fieldMask: z.string().trim(),
-            style: z.object({
-                color: z.string().optional(),
-                icon: z.string().optional(),
-            }),
-            domainType: z.union([z.literal('AGGREGATE'), z.literal('TRACKER')]),
-            valueType: z.string(),
-            aggregationType: z.string(),
-            optionSet: z.object({ id: z.string() }),
-            commentOptionSet: z.object({ id: z.string() }),
-            legendSets: z.array(z.object({ id: z.string() })),
-            aggregationLevels: z.array(z.number()),
-            attributeValues: z.array(
-                z.object({
-                    value: z.string().optional(),
-                    attribute: z.object({
-                        id: z.string(),
-                    }),
-                })
-            ),
+            color: z.string().optional(),
+            icon: z.string().optional(),
         })
-        .partial()
-)
+        .default({}),
+    domainType: z
+        .nativeEnum(DataElement.domainType)
+        .default(DataElement.domainType.AGGREGATE),
+    valueType: z
+        .nativeEnum(DataElement.valueType)
+        .default(DataElement.valueType.TEXT),
+    aggregationType: z
+        .nativeEnum(DataElement.aggregationType)
+        .default(DataElement.aggregationType.NONE),
+    categoryCombo: z
+        .object({ id: z.string(), displayName: z.string().optional() })
+        .default(DEFAULT_CATEGORY_COMBO),
+    optionSet: z
+        .object({ id: z.string(), displayName: z.string().optional() })
+        .optional(),
+    commentOptionSet: z
+        .object({ id: z.string(), displayName: z.string().optional() })
+        .optional(),
+    legendSets: z
+        .array(z.object({ id: z.string(), displayName: z.string().optional() }))
+        .default([]),
+    aggregationLevels: z.array(z.number()).default([]),
+    zeroIsSignificant: z.boolean().default(false),
+})
+export const dataElementListSchema = dataElementBaseSchema
+    .merge(withDefaultListColumns)
+    .merge(withAttributeValues)
+    .extend({
+        name: z.string(),
+    })
+
+export const dataElementFormSchema = dataElementBaseSchema
+    .merge(identifiable)
+    .merge(withAttributeValues)
+
+export const initialValues = getDefaults(dataElementFormSchema)
+
+export type DataElementFormValues = typeof initialValues
+
+export const validate = createFormValidate(dataElementFormSchema)
+
+export const dataElementValueFormatter = (values: DataElementFormValues) => ({
+    ...values,
+    attributeValues: values.attributeValues.filter(({ value }) => !!value),
+    categoryCombo: values.categoryCombo.id ? values.categoryCombo : undefined,
+    commentOptionSet: values.commentOptionSet?.id
+        ? values.commentOptionSet
+        : undefined,
+    optionSet: values.optionSet?.id ? values.optionSet : undefined,
+})