+++ src/pages/programIndicatorGroups/form/index.ts
@@ -1,6 +1,6 @@
 export {
-    ProgramIndicatorGroupFormSchema,
-    ProgramIndicatorGroupListSchema,
+    programIndicatorGroupFormSchema,
+    programIndicatorGroupListSchema,
     initialValues,
     validate,
 } from './ProgramIndicatorGroupsSchema'