+++ src/pages/programIndicators/form/programIndicatorsFormSchema.ts
@@ -10,49 +10,45 @@ const {
     modelReference,
     withAttributeValues,
     withDefaultListColumns,
+    style,
 } = modelFormSchemas
 
-const ProgramIndicatorsBaseSchema = z.object({
+const programIndicatorsBaseSchema = z.object({
     program: modelReference,
     aggregationType: z.nativeEnum(ProgramIndicator.aggregationType).optional(),
     analyticsType: z
         .nativeEnum(ProgramIndicator.analyticsType)
         .default(ProgramIndicator.analyticsType.EVENT),
     displayInForm: z.boolean().default(false),
+    description: z.string().optional(),
     legendSets: referenceCollection.default([]),
     aggregateExportCategoryOptionCombo: z.string().optional(),
     aggregateExportAttributeOptionCombo: z.string().optional(),
+    aggregateExportDataElement: z.string().optional(),
     expression: z.string().optional(),
     filter: z.string().optional(),
     orgUnitField: z.string().optional(),
     decimals: z.number().int().lte(5).gte(0).optional(),
+    code: z.string().optional(),
 })
 
-export const ProgramIndicatorsListSchema = ProgramIndicatorsBaseSchema.merge(
-    withDefaultListColumns
-)
+export const programIndicatorsListSchema = programIndicatorsBaseSchema
+    .merge(withDefaultListColumns)
     .merge(withAttributeValues)
     .extend({
-        program: modelReference,
         displayShortName: z.string(),
         displayDescription: z.string().optional(),
         displayFormName: z.string().optional(),
         user: UserSchema,
         favorite: z.boolean(),
-        code: z.string().optional(),
     })
 
-export const programIndicatorFormSchema = ProgramIndicatorsBaseSchema.merge(
-    identifiable
-)
+export const programIndicatorFormSchema = programIndicatorsBaseSchema
+    .merge(identifiable)
     .merge(withAttributeValues)
     .extend({
         shortName: z.string().trim(),
-        code: z.string().trim().optional(),
-        style: z.object({
-            color: z.string().optional(),
-            icon: z.string().optional(),
-        }),
+        style: style.optional(),
         orgUnitField: z.string().optional(),
         analyticsPeriodBoundaries: z
             .array(